Implementation of instance and instance header (closed #1)
This commit closes #1 by adding the implementations of Instance and
Instance Header to the project, as well as DenseInstance.

The modules where implemented as follows:

Instance: Trait used by DenseInstance to define it's methods

InstanceHeader: Struct containing the attributes and relation names, as
well as class index for the instance.

DenseInstance: Struct containing the concrete implementation of
Instance.

These implementation currently follows the definitions in the project's
class diagram and are subject to changes in the future as needed.