Implementation of Stream trait (closed #3)
This commit closes #3 by implementing the Stream trait for instance
management. The trait was implemented with the following methods:

header(): returns the instance header of the stream, to know attibutes
and classes.

has_more_instances(): returns a boolean indicating if there are more
instances in the stream.

next_instance(): returns an option for a Box<> of the next instance,
since instance is a trait implemented concretly by DenseInstance.

This implementation currently follows the definitions in the project's
class diagram and are subject to changes in the future as needed