Implementation of incomplete classes methods and attributes
This commit closes #13 by implementing the remaining methods missing from the first implementations of Attribute (Nominal and Numeric), Instance, InstanceHeader and Stream.

Instance:
 - set_weight()
 - set_value_at_index()
 - is_missing_at_index()
 - attribute_at_index()
 - index_of_attribute()
 - class_index()
 - set_class_value()
 - is_class_missing()
 - number_of_classes()
 - to_vec()

InstanceHeader:
 - relation_name()
 - attribute_at_index()
 - index_of_attribute()
 - class_index()
 - number_of_classes()

Attribute:
 - arff_representation()

Some aditional methods where implemented as well, like as_any() for numeric and nominal attributes, used to downreference an Attribute.

The implementation from this methods are subject to change in the future as needed.