ci: temporarily disabled failure on warnings
Since the project is not in a stable state at this point, there will be
lots of warnings. This caused the pipeline to not provide any useful
information, since warnings would cause it to fail constantly.

This commit disables failures on warnings, which will make the pipeline
useful once again, at least telling us if the tests are passing.