feat: Implementation of AttributeClassObserver
This commit closes #21 by implementing the AttributeClassObserver trait and it's concrete implementation NominalAttributeClassObserver and GaussianNumericAttributeClassObserver.

These classes are essential for the future implementation of classifiers, as they are responsible for keep statistics for each attribute conditioned to a class.

This commits also include the implementation of GaussianEstimator, essential for the GaussianNumericAttributeClassObserver