feat: Implemeted nodes for Hoeffding Tree
this commit closes #25 by implementing the Node structure and all it's related structs. The implemented modules were:
 - Node
 - FoundNode
 - SplitNode
 - LearningNode
 - ActiveLearningNode
 - InactiveLearningNode
 - LearningNodeNB
 - LearningNodeNBAdaptive

Some details to take note about this implementation:

The method filter_instance_to_leaf in SplitNode was implemented using static lifetimes.

Some parts of the hoeffding tree was implemented, due to node methods depending on them.

The new_nominal and new_numeric_class_observer from hoeffding tree were implemented as if the PreparedClassOption from moa were disabled.

Tests of these implementations are not yet implemented, as they depend on the implementation of the HoeffdingTree