feat: Implemeted nodes for Hoeffding Tree (closed #25)
this commit closes #25 by implementing the Node structure and all it's
related structs. The implemented modules were:
- Node
- FoundNode
- SplitNode
- LearningNode
- ActiveLearningNode
- InactiveLearningNode
- LearningNodeNB
- LearningNodeNBAdaptive

Some details to take note about this implementation:

The method `filter_instance_to_leaf()` in SplitNode was implemented
using static lifetimes.

Some parts of the hoeffding tree was implemented, due to node methods
depending on them.

The `new_nominal_class_observer()` and `new_numeric_class_observer()`
from hoeffding tree were implemented as if the PreparedClassOption from
moa was disabled.

Tests of these implementations are not yet implemented, as they depend
on the implementation of the HoeffdingTree

In the Classifier Trait, the methos get_votes_for_instance was changed
to return only an `Vec<f64>` instead of an `Option<Vec<f64>>`