feat: Implementation of Hoeffding Trees (closed #18)
This commit closes #18 by implementing the Hoeffding Tree struct. This
addition brings a lot of changes to te project, as some classes nedded
adaptations to attent to the tree implementation.

Options where added to the constructor of the struct, and need to be
validated and defined by the user.

Other classes needed to implement the training and classification where
also added, such as instance conditional test and null attribute class
observer

Tests were implemented for HoeffdingTree, but need to be added in the
other classes